+++ dataSources/anemone/db/config.json
@@ -6,8 +6,8 @@
     "regexMatch": ".*\\.tsv\\.xz",
     "perFileProcessing": [
         {
-            "path": "lib/tools/extractor.py",
-            "function": "Extractor.extract",
+            "path": "lib/tools/zipping.py",
+            "function": "extract",
             "args": [
                 "{INPUT, 0}",
                 "{PATH, PROCESSING}"
