+++ dataSources/bold/datapackage/config.json
@@ -12,8 +12,8 @@
     },
     "perFileProcessing": [
         {
-            "path": "lib/tools/extractor.py",
-            "function": "Extractor.extract",
+            "path": "lib/tools/zipping.py",
+            "function": "extract",
             "args": [
                 "{INPUT, 0, FILEPATH}",
                 "{PATH, PROCESSING}"
