+++ dataSources/ncbi/nucleotide/config.json
@@ -5,8 +5,8 @@
     "regexMatch": "gb[a-z]{3}\\d+\\.seq\\.gz",
     "perFileProcessing": [
         {
-            "path": "lib/tools/extractor.py",
-            "function": "Extractor.extract",
+            "path": "lib/tools/zipping.py",
+            "function": "extract",
             "args": [
                 "{INPUT, 0, FILEPATH}",
                 "{PATH, PROCESSING}"
