+++ dataSources/ncbi/taxonomy/config.json
@@ -6,8 +6,8 @@
             "downloadedFile": "taxdump.zip",
             "processing": [
                 {
-                    "path": "lib/tools/extractor.py",
-                    "function": "Extractor.extract",
+                    "path": "lib/tools/zipping.py",
+                    "function": "extract",
                     "args": [
                         "{INPUT, 0, FILEPATH}",
                         "{PATH, PROCESSING}"
