+++ dataSources/ncbi/tpa/config.json
@@ -6,8 +6,8 @@
             "downloadedFile": "con_tpa_cu.gbff.gz",
             "processing": [
                 {
-                    "path": "lib/tools/extractor.py",
-                    "function": "Extractor.extract",
+                    "path": "lib/tools/zipping.py",
+                    "function": "extract",
                     "args": [
                         "{INPUT, 0, FILEPATH}",
                         "{PATH, PROCESSING}"
@@ -33,8 +33,8 @@
             "downloadedFile": "tpa_cu.gbff.gz",
             "processing": [
                 {
-                    "path": "lib/tools/extractor.py",
-                    "function": "Extractor.extract",
+                    "path": "lib/tools/zipping.py",
+                    "function": "extract",
                     "args": [
                         "{INPUT, 0, FILEPATH}"
                     ],
