+++ dataSources/silva/metadata/config.json
@@ -6,8 +6,8 @@
     "subDirectoryDepth": 0,
     "perFileProcessing": [
         {
-            "path": "lib/tools/extractor.py",
-            "function": "Extractor.extract",
+            "path": "lib/tools/zipping.py",
+            "function": "extract",
             "args": [
                 "{INPUT, 0, FILEPATH}",
                 "{PATH, PROCESSING}"
